python/requirements.txt
azure-identity
azure-servicebus
azure-storage-queue
python-dotenv